Save and download generated pmdl file
