+++ Susi/Model/Client/Convenience.swift
@@ -300,7 +300,7 @@ extension Client {
 
     // MARK: - Train hotword using Snowboy API
 
-    func trainHotwordUsingSnowboy(_ params: [String : AnyObject], _ completion: @escaping(_ response: Any?, _ success: Bool, _ error: String?) -> Void) {
+    func trainHotwordUsingSnowboy(_ params: [String : AnyObject], _ completion: @escaping(_ success: Bool, _ error: String?) -> Void) {
 
         let urlString = getApiUrl(APIURLs.SnowboyTrain)
 
@@ -308,17 +308,35 @@ extension Client {
         var request = URLRequest(url: url)
         request.httpMethod = "POST"
         request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-                
+
         request.httpBody = try! JSONSerialization.data(withJSONObject: params)
 
         Alamofire.request(request).responseData { (results) in
-            print(results.error)
-            print(results)
-            print(results.response)
-            print(results.result)
-            print(results.response?.statusCode)
+            if results.response?.statusCode == 201 {
+                if let data = results.data {
+                    //this is the file. we will write to and read from it
+                    let file = "susi.pmdl"
+
+                    if let dir = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first {
+
+                        let path = dir.appendingPathComponent(file)
+
+                        //writing
+                        do {
+                            try data.write(to: path)
+                            completion(true, nil)
+                        } catch let error {
+                            print(error.localizedDescription)
+                            completion(false, ResponseMessages.ServerError)
+                        }
+                    }
+                }
+            } else {
+                print(results.error?.localizedDescription ?? "Error")
+                completion(false, ResponseMessages.ServerError)
+            }
         }
-        
+
     }
 
 }
